LayoutTests/imported/w3c/web-platform-tests/intersection-observer/shadow-content-expected.txt


PASS Observing a target inside shadow DOM. 
FAIL First rAF after creating shadow DOM. assert_equals: entries.length expected 1 but got 0
